+++ experimental/traffic-portal/src/app/shared/interceptor/alerts.interceptor.ts
@@ -44,7 +44,7 @@ export class AlertInterceptor implements HttpInterceptor {
 				if (Object.prototype.hasOwnProperty.call(r, "body") &&
 				    Object.prototype.hasOwnProperty.call((r as {body: unknown}).body, "alerts")) {
 					for (const a of (r as {body: {alerts: Array<unknown>}}).body.alerts) {
-						this.alertService.alertsSubject.next(a as Alert);
+						this.alertService.newAlert(a as Alert);
 					}
 				}
 			}